pyproject.toml
[tool.poetry]
name = "data whisper - conversational bi"
version = "0.1.0"
description = "Conversational BI"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.43.0"
scikit-learn = "^1.5.1"
seaborn = "^0.13.2"
plotly = "^5.24.0"
pandas = "^2.2.2"
streamlit = "^1.38.0"
streamlit-extras = "^0.4.7"
streamlit-aggrid = "^1.0.5"
streamlit-ace = "^0.1.1"
streamlit-pandas-profiling = "^0.1.3"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"
tabulate = "^0.9.0"
python-dotenv = "^1.0.1"
langgraph = "^0.2.16"
langchainhub = "^0.1.21"
langchain-experimental = "^0.0.65"
ipython = "^8.27.0"
reportlab = "^4.2.2"
fpdf = "^1.7.2"
kaleido = "0.2.1"
ipykernel = "^6.29.5"
nbformat = "^5.10.4"
python-docx = "^1.1.2"
black = "^24.8.0"
faiss-cpu = "^1.9.0.post1"
asyncio = "^3.4.3"
sqlalchemy = "^2.0.36"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.34.0"
psycopg2 = "^2.9.10"
psycopg-pool = "^3.2.4"
psycopg = {extras = ["binary"], version = "^3.2.3"}
langgraph-checkpoint-postgres = "^2.0.10"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
